requirements.txt
scikit-learn
matplotlib
torch===1.4.0
torchvision===0.5.0
torchsummary
Pillow
scikit-image
liac-arff
flake8